improving the implementation of the AbstractSmell class, removing repeated methods in the child classes, simplifying the code.
