+++ src/main/java/testsmell/smell/AssertionRoulette.java
@@ -20,26 +20,12 @@ import java.util.List;
  */
 public class AssertionRoulette extends AbstractSmell {
 
-    private List<SmellyElement> smellyElementList;
-
-    public AssertionRoulette() {
-        smellyElementList = new ArrayList<>();
-    }
-
     /**
      * Checks of 'Assertion Roulette' smell
      */
     @Override
     public String getSmellName() {
         return "Assertion Roulette";
-    }
-
-    /**
-     * Returns true if any of the elements has a smell
-     */
-    @Override
-    public boolean getHasSmell() {
-        return smellyElementList.stream().filter(x -> x.getHasSmell()).count() >= 1;
     }
 
     /**
@@ -51,15 +37,6 @@ public class AssertionRoulette extends AbstractSmell {
         classVisitor = new AssertionRoulette.ClassVisitor();
         classVisitor.visit(testFileCompilationUnit, null);
     }
-
-    /**
-     * Returns the set of analyzed elements (i.e. test methods)
-     */
-    @Override
-    public List<SmellyElement> getSmellyElements() {
-        return smellyElementList;
-    }
-
 
     private class ClassVisitor extends VoidVisitorAdapter<Void> {
         private MethodDeclaration currentMethod = null;
