+++ src/main/java/testsmell/smell/ConstructorInitialization.java
@@ -20,12 +20,7 @@ If this code detects the existence of a constructor, it sets the class as smelly
  */
 public class ConstructorInitialization extends AbstractSmell {
 
-    private List<SmellyElement> smellyElementList;
     private String testFileName;
-
-    public ConstructorInitialization() {
-        smellyElementList = new ArrayList<>();
-    }
 
     /**
      * Checks of 'Constructor Initialization' smell
@@ -33,14 +28,6 @@ public class ConstructorInitialization extends AbstractSmell {
     @Override
     public String getSmellName() {
         return "Constructor Initialization";
-    }
-
-    /**
-     * Returns true if any of the elements has a smell
-     */
-    @Override
-    public boolean getHasSmell() {
-        return smellyElementList.stream().filter(x -> x.getHasSmell()).count() >= 1;
     }
 
     /**
@@ -53,15 +40,6 @@ public class ConstructorInitialization extends AbstractSmell {
         classVisitor = new ConstructorInitialization.ClassVisitor();
         classVisitor.visit(testFileCompilationUnit, null);
     }
-
-    /**
-     * Returns the set of analyzed elements (i.e. test methods)
-     */
-    @Override
-    public List<SmellyElement> getSmellyElements() {
-        return smellyElementList;
-    }
-
 
     private class ClassVisitor extends VoidVisitorAdapter<Void> {
         TestClass testClass;
