+++ src/main/java/testsmell/smell/DuplicateAssert.java
@@ -14,26 +14,12 @@ import java.util.*;
 
 public class DuplicateAssert extends AbstractSmell {
 
-    private List<SmellyElement> smellyElementList;
-
-    public DuplicateAssert() {
-        smellyElementList = new ArrayList<>();
-    }
-
     /**
      * Checks of 'Duplicate Assert' smell
      */
     @Override
     public String getSmellName() {
         return "Duplicate Assert";
-    }
-
-    /**
-     * Returns true if any of the elements has a smell
-     */
-    @Override
-    public boolean getHasSmell() {
-        return smellyElementList.stream().filter(x -> x.getHasSmell()).count() >= 1;
     }
 
     /**
@@ -45,15 +31,6 @@ public class DuplicateAssert extends AbstractSmell {
         classVisitor = new DuplicateAssert.ClassVisitor();
         classVisitor.visit(testFileCompilationUnit, null);
     }
-
-    /**
-     * Returns the set of analyzed elements (i.e. test methods)
-     */
-    @Override
-    public List<SmellyElement> getSmellyElements() {
-        return smellyElementList;
-    }
-
 
     private class ClassVisitor extends VoidVisitorAdapter<Void> {
         private MethodDeclaration currentMethod = null;
