+++ src/main/java/testsmell/smell/EmptyTest.java
@@ -18,26 +18,12 @@ import java.util.List;
  */
 public class EmptyTest extends AbstractSmell {
 
-    private List<SmellyElement> smellyElementList;
-
-    public EmptyTest() {
-        smellyElementList = new ArrayList<>();
-    }
-
     /**
      * Checks of 'Empty Test' smell
      */
     @Override
     public String getSmellName() {
         return "EmptyTest";
-    }
-
-    /**
-     * Returns true if any of the elements has a smell
-     */
-    @Override
-    public boolean getHasSmell() {
-        return smellyElementList.stream().filter(x -> x.getHasSmell()).count() >= 1;
     }
 
     /**
@@ -48,14 +34,6 @@ public class EmptyTest extends AbstractSmell {
         EmptyTest.ClassVisitor classVisitor;
         classVisitor = new EmptyTest.ClassVisitor();
         classVisitor.visit(testFileCompilationUnit, null);
-    }
-
-    /**
-     * Returns the set of analyzed elements (i.e. test methods)
-     */
-    @Override
-    public List<SmellyElement> getSmellyElements() {
-        return smellyElementList;
     }
 
     /**
