+++ src/main/java/testsmell/smell/LazyTest.java
@@ -25,13 +25,12 @@ public class LazyTest extends AbstractSmell {
     private static final String TEST_FILE = "Test";
     private static final String PRODUCTION_FILE = "Production";
     private String productionClassName;
-    private List<SmellyElement> smellyElementList;
     private List<MethodUsage> calledProductionMethods;
     private List<MethodDeclaration> productionMethods;
 
     public LazyTest() {
+        super();
         productionMethods = new ArrayList<>();
-        smellyElementList = new ArrayList<>();
         calledProductionMethods = new ArrayList<>();
     }
 
@@ -41,14 +40,6 @@ public class LazyTest extends AbstractSmell {
     @Override
     public String getSmellName() {
         return "Lazy Test";
-    }
-
-    /**
-     * Returns true if any of the elements has a smell
-     */
-    @Override
-    public boolean getHasSmell() {
-        return smellyElementList.stream().filter(x -> x.getHasSmell()).count() >= 1;
     }
 
     /**
@@ -80,14 +71,6 @@ public class LazyTest extends AbstractSmell {
                 }
             }
         }
-    }
-
-    /**
-     * Returns the set of analyzed elements (i.e. test methods)
-     */
-    @Override
-    public List<SmellyElement> getSmellyElements() {
-        return smellyElementList;
     }
 
     private class MethodUsage {
