+++ src/main/java/testsmell/smell/ResourceOptimism.java
@@ -17,26 +17,12 @@ import java.util.List;
 
 public class ResourceOptimism extends AbstractSmell {
 
-    private List<SmellyElement> smellyElementList;
-
-    public ResourceOptimism() {
-        smellyElementList = new ArrayList<>();
-    }
-
     /**
      * Checks of 'Resource Optimism' smell
      */
     @Override
     public String getSmellName() {
         return "Resource Optimism";
-    }
-
-    /**
-     * Returns true if any of the elements has a smell
-     */
-    @Override
-    public boolean getHasSmell() {
-        return smellyElementList.stream().filter(x -> x.getHasSmell()).count() >= 1;
     }
 
     /**
@@ -48,15 +34,6 @@ public class ResourceOptimism extends AbstractSmell {
         classVisitor = new ResourceOptimism.ClassVisitor();
         classVisitor.visit(testFileCompilationUnit, null);
     }
-
-    /**
-     * Returns the set of analyzed elements (i.e. test methods)
-     */
-    @Override
-    public List<SmellyElement> getSmellyElements() {
-        return smellyElementList;
-    }
-
 
     private class ClassVisitor extends VoidVisitorAdapter<Void> {
         private MethodDeclaration currentMethod = null;
