+++ src/main/java/testsmell/smell/UnknownTest.java
@@ -19,26 +19,12 @@ import java.util.Optional;
 
 public class UnknownTest extends AbstractSmell {
 
-    private List<SmellyElement> smellyElementList;
-
-    public UnknownTest() {
-        smellyElementList = new ArrayList<>();
-    }
-
     /**
      * Checks of 'Unknown Test' smell
      */
     @Override
     public String getSmellName() {
         return "Unknown Test";
-    }
-
-    /**
-     * Returns true if any of the elements has a smell
-     */
-    @Override
-    public boolean getHasSmell() {
-        return smellyElementList.stream().filter(x -> x.getHasSmell()).count() >= 1;
     }
 
     /**
@@ -50,15 +36,6 @@ public class UnknownTest extends AbstractSmell {
         classVisitor = new UnknownTest.ClassVisitor();
         classVisitor.visit(testFileCompilationUnit, null);
     }
-
-    /**
-     * Returns the set of analyzed elements (i.e. test methods)
-     */
-    @Override
-    public List<SmellyElement> getSmellyElements() {
-        return smellyElementList;
-    }
-
 
     private class ClassVisitor extends VoidVisitorAdapter<Void> {
         private MethodDeclaration currentMethod = null;
