+++ src/main/java/testsmell/smell/VerboseTest.java
@@ -17,26 +17,12 @@ If a test methods contains a statements that exceeds a certain threshold, the me
  */
 public class VerboseTest extends AbstractSmell {
 
-    private List<SmellyElement> smellyElementList;
-
-    public VerboseTest() {
-        smellyElementList = new ArrayList<>();
-    }
-
     /**
      * Checks of 'Verbose Test' smell
      */
     @Override
     public String getSmellName() {
         return "Verbose Test";
-    }
-
-    /**
-     * Returns true if any of the elements has a smell
-     */
-    @Override
-    public boolean getHasSmell() {
-        return smellyElementList.stream().filter(x -> x.getHasSmell()).count() >= 1;
     }
 
     /**
@@ -47,14 +33,6 @@ public class VerboseTest extends AbstractSmell {
         VerboseTest.ClassVisitor classVisitor;
         classVisitor = new VerboseTest.ClassVisitor();
         classVisitor.visit(testFileCompilationUnit, null);
-    }
-
-    /**
-     * Returns the set of analyzed elements (i.e. test methods)
-     */
-    @Override
-    public List<SmellyElement> getSmellyElements() {
-        return smellyElementList;
     }
 
     private class ClassVisitor extends VoidVisitorAdapter<Void> {
